==> 1_data_preparation.ipynb <==
import warnings
warnings.filterwarnings('ignore')

import pandas as pd
import numpy as np
import os.path as path

df_train = pd.read_csv(path.join('1_rawdata', 'application_train.csv'))
df_test = pd.read_csv(path.join("1_rawdata", "application_test.csv"))

# concat before dummies
df = pd.concat([df_train, df_test], axis="index")

# easy data preparation by dropping all nulls
df = df.dropna(axis='index')

# getting y
df_y = pd.DataFrame(df.TARGET)

# getting X
df.drop(df[['TARGET']], axis="columns",inplace=True)
df_X = pd.get_dummies(df)

# verifying shapes
print("df_X : ", df_X.shape, "\ndf_y : ", df_y.shape)

# store x and y
df_X.to_csv(path.join("2_dataprep", "prep_train_X.csv"), index=False)
df_y.to_csv(path.join("2_dataprep", "prep_train_y.csv"), index=False)
