==> 2_feature_engineering.ipynb <==
import warnings
warnings.filterwarnings('ignore')

import pandas as pd
import os.path as path
from pickle import dump

from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler

df_X = pd.read_csv(path.join("2_dataprep", "prep_train_X.csv"))
df_y = pd.read_csv(path.join("2_dataprep", "prep_train_y.csv"))

# splitting here because of the dummies done before, if we reverse the algo, we may have troubles with shapes and columns 
X_train, X_test, y_train, y_test = train_test_split(df_X, df_y, test_size=0.33, random_state=42)

# fit and transform X train
scaler = StandardScaler()
scaler.fit(df_X)
X_train_scaled = scaler.transform(X_train)

# we use the fit method from standarscaler, only for train dataframe, not for test
# transform X test
X_test_scaled = scaler.transform(X_test)

# create scaled dataframes 
X_train_scaled = pd.DataFrame(data=X_train_scaled, columns=X_train.columns)
X_test_scaled = pd.DataFrame(data=X_test_scaled, columns=X_test.columns)

# save X traind and test, scaled dataframes  
X_train_scaled.to_csv(path.join("3_fitted_dataframe", "X_train_scaled.csv"), index=False)
X_test_scaled.to_csv(path.join("3_fitted_dataframe", "X_test_scaled.csv"), index=False)
# save y train and test dataframes
y_train.to_csv(path.join("3_5_y_dataframe", "y_train.csv"), index=False)
y_test.to_csv(path.join("3_5_y_dataframe", "y_test.csv"), index=False)

# save scaler in case of
dump(scaler, open((path.join('4_fitted_scaler', 'scaler.pkl')), 'wb'))
