==> 3_model_training.ipynb <==
import warnings
warnings.filterwarnings('ignore')

import pandas as pd 
from pickle import load
import os.path as path
from pickle import load, dump

from sklearn.ensemble import RandomForestClassifier
from sklearn.ensemble import GradientBoostingClassifier
from xgboost import XGBClassifier

# load dataframes
X_train_scaled = pd.read_csv(path.join("3_X_fitted_dataframe", "X_train_scaled.csv"))
y_train = pd.read_csv(path.join('4_y_dataframe', 'y_train.csv'))
X_test_scaled = pd.read_csv(path.join("3_X_fitted_dataframe", "X_test_scaled.csv"))
y_test = pd.read_csv(path.join('4_y_dataframe', 'y_test.csv'))

# fit xgboot model
XGB_clf = XGBClassifier()
XGB_clf.fit(X_train_scaled, y_train)

# fit random forest model
rf_clf = RandomForestClassifier()
rf_clf.fit(X_train_scaled, y_train)

# fit Gradient boostinf model
GBoost_clf = GradientBoostingClassifier()
GBoost_clf.fit(X_train_scaled, y_train)

# save models
dump(rf_clf, open(path.join('6_models', 'randomForest.pkl'), 'wb'))
dump(XGB_clf, open(path.join('6_models', 'xgBoost.pkl'), 'wb'))
dump(GBoost_clf, open(path.join('6_models', 'GradientBoosting.pkl'), 'wb'))
