==> 4_predict.ipynb <==
import warnings
warnings.filterwarnings('ignore')

import pandas as pd 
from pickle import load
import os.path as path
from pickle import load, dump

from sklearn.ensemble import RandomForestClassifier
from sklearn.ensemble import GradientBoostingClassifier
from xgboost import XGBClassifier

# load dataframes
X_test_scaled = pd.read_csv(path.join("3_X_fitted_dataframe", "X_test_scaled.csv"))
y_test = pd.read_csv(path.join("4_y_dataframe", "y_test.csv"))

# load models
rf_clf = load(open(path.join('6_models', 'randomforest.pkl'), 'rb'))
XGB_clf = load(open(path.join('6_models', 'xgBoost.pkl'), 'rb'))
GBoost_clf = load(open(path.join('6_models', 'GradientBoosting.pkl'), 'rb'))

# score with xgBoost
score_xgb = XGB_clf.score(X_test_scaled, y_test)
print(score_xgb)

# score with GradientBoosting
score_gb = GBoost_clf.score(X_test_scaled, y_test)
print(score_gb)

# score with Random Forest
score_rf = rf_clf.score(X_test_scaled, y_test)
print(score_rf)

# predict with xgBoost
predXGB = XGB_clf.predict(X_test_scaled)

# predict with GradientBoosting
predGBoost = GBoost_clf.predict(X_test_scaled)

# predict with Random Forest
predRF = rf_clf.predict(X_test_scaled)
